# target/document-storage-rest-api-0.0.1-SNAPSHOT/WEB-INF/classes/static/swagger-apis/api1/swagger.yaml
swagger: "2.0"
info:
  description: "Document Storage API Service"
  version: "1.0.0"
  title: "Document Storage  Rest API Service."
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "sandeeprgurav@gmail.com"
tags:
  - name: "document"
    description: "API for Document CURD Operations. Using this API we can upload any kind of documents."    
schemes:  
  - "http"
paths:
  /storage/documents:
    post:
      tags:
        - "document"
      summary: "Upload Document API."
      description: "API for Document Upload into server."
      consumes:
        - "multipart/form-data"
      produces:        
        - "application/json"
      parameters:
      - name: "document"
        in: "formData"
        description: "Upload Document"
        required: true
        type: "file"
      responses: 
        201: 
          description: "Document uploaded Successfully!"
          schema: 
            type: "object"
        204: 
          description: "No Document uploaded!"
          schema: 
            type: "object"                  
  /storage/documents/{documentId}:
    get:
      tags:
        - "document"
      summary: "Download Document API"
      description: "API to download document by passing document Id."
      consumes:
        - "application/json"      
      parameters:
      - name: "documentId"
        in: "path"
        description: "Document ID"
        required: true
        type: "string"
      responses: 
        200: 
          description: "Found document for entered documentId"
          schema: 
            type: "object"
        404: 
          description: "Invalid Document Id entered."
          schema: 
            type: "object"               
  /storage/documents/{id}:
    put:
      tags:
        - "document"
      summary: "Document Update API"
      description: "API for Document Update using document Id."
      consumes:
        - "multipart/form-data"      
      parameters:
      - name: "id"
        in: "path"
        description: "Document ID"
        required: true
        type: "string"
      - name: "document"
        in: "formData"
        description: "Upload Document"
        required: true
        type: "file"        
      responses: 
        204: 
          description: "Document Updated for passed document Id"
          schema: 
            type: "object"                  
        404: 
          description: "Invalid Document Id entered."
          schema: 
            type: "object"                              
  /storage/documents/{deleteDocumentId}:
    delete:
      tags:
        - "document"
      summary: "Document Delete API."
      description: "API for Document delete from the server."
      consumes:
        - "application/json"      
      parameters:
      - name: "deleteDocumentId"
        in: "path"
        description: "Document ID"
        required: true
        type: "string"
      responses: 
        204: 
          description: "Document deleted successfully!"
          schema: 
            type: "object"                  
        404: 
          description: "Invalid Document Id entered."
          schema: 
            type: "object"                                          